yapayzekaogretmen_python/backend/requirements_modern.txt
# AI Öğretmen (aiogretmen.com) - Modern Requirements
# Python 3.11+ Compatible

# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0
redis[hiredis]==5.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0

# AI/ML
openai==1.8.0
langchain==0.1.0
langchain-community==0.0.12
langchain-openai==0.0.5
chromadb==0.4.22
tiktoken==0.5.2
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Search & Analytics
elasticsearch==8.11.1
faiss-cpu==1.7.4

# Document Processing
pypdf==3.17.4
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0

# Communication
twilio==8.11.0
sendgrid==6.11.0
python-telegram-bot==20.7

# Real-time & Background Tasks
websockets==12.0
celery[redis]==5.3.4
python-socketio==5.10.0

# GraphQL
strawberry-graphql[fastapi]==0.217.1

# Monitoring & Logging
loguru==0.7.2
sentry-sdk[fastapi]==1.39.1
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# AWS Services (Optional)
boto3==1.34.14
aioboto3==12.2.0

# Blockchain (Optional)
web3==6.14.0
eth-account==0.10.0

# Additional ML Libraries
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1

# Voice & Audio
pydub==0.25.1
SpeechRecognition==3.10.1
gTTS==2.5.0

# Computer Vision
opencv-python==4.9.0.80
pytesseract==0.3.10

# Notifications
firebase-admin==6.4.0

# API Documentation
fastapi-pagination==0.12.13
fastapi-limiter==0.1.5

# CORS
fastapi-cors==0.0.6

# Email
aiosmtplib==3.0.1
email-validator==2.1.0.post1

# Scheduling
apscheduler==3.10.4

# Data Validation
marshmallow==3.20.1
pydantic-extra-types==2.4.1

# Caching
aiocache[redis]==0.12.2

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
arrow==1.3.0
humanize==4.9.0
python-slugify==8.0.1
